hw L1.3 channel
