+++ L1.1.go
@@ -1,8 +1,7 @@
-package model
+package main
 
-import "fmt" // подключение стандартного пакета "fmt" для печати в консоль
+import "fmt"
 
-// Определение структуры Human
 type Human struct {
 	Name       string // Имя человека
 	NameDative string // Имя человека в дательном падеже
@@ -71,4 +70,29 @@ func (a Action) ShowActivity() {
 func (a *Action) StopActivity() {
 	fmt.Println(a.Name, "закончил(а) заниматься", a.Activity) // останавливаем активность и выводим сообщение
 	a.Activity = ""
+}
+
+func Person() {
+	// Создаём объект Action с ID=1 и встроенным Human
+	a := Action{
+		ID: 1,
+		Human: Human{
+			Name:       "Анастасия",
+			NameDative: "Анастасии",
+			Age:        29,
+			Gender:     "Женщина",
+			Phone:      "555-555-5555",
+			Email:      "anastasia@gmail.com",
+			Address:    "улица Ленина дом 1",
+		},
+	}
+
+	a.SayHello()               // вызываем метод Human через embedding, здороваемся
+	a.ContactInfo()            // выводим контактные данные
+	a.Eat("апельсин")          // человек ест мясо
+	a.Birthday()               // увеличиваем возраст на 1
+	a.Sleep(8)                 // показываем, что человек спит 8 часов
+	a.StartActivity("jogging") // начинаем бегать трусцой
+	a.ShowActivity()           // Демонстрируем текущую активность при наличии
+	a.StopActivity()           // останавливаемся
 }
